Moved controls into their own namespace
