+++ AllReadyApp/AllReady.BrowserTest/Controls/DateTimePicker.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace AllReady.BrowserTest.Pages
+namespace AllReady.BrowserTest.Controls
 {
     class DateTimePicker : IControl
     {
