+++ AllReadyApp/AllReady.BrowserTest/Controls/TinyMCE.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace AllReady.BrowserTest.Pages
+namespace AllReady.BrowserTest.Controls
 {
     class TinyMCE : IControl
     {
