+++ AllReadyApp/AllReady.BrowserTest/Pages/AdminCampaignCreatePage.cs
@@ -1,3 +1,4 @@
+using AllReady.BrowserTest.Controls;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Interactions;
 using OpenQA.Selenium.Support.UI;
