+++ AllReadyApp/AllReady.BrowserTest/Pages/AdminEventCreatePage.cs
@@ -1,3 +1,4 @@
+using AllReady.BrowserTest.Controls;
 using OpenQA.Selenium;
 using System;
 using System.Collections.Generic;
